Code in Java
Polymorphism is a key pillar of Object-Oriented Programming (OOP) that enables objects of different classes to be treated as objects of a common superclass. This allows various classes to share a common interface or method name while having their specific implementations.

There are two types of polymorphism:
1) Compile-time Polymorphism (Method Overloading):
It occurs when two or more methods in the same class have the same name but different parameters. The compiler determines which method to call based on the number or type of arguments provided during the method call.

2) Runtime Polymorphism (Method Overriding):
It occurs when a subclass provides a specific implementation for a method that is already defined in its superclass. During runtime, the method of the subclass is called based on the actual object type, even if the reference is of the superclass type.