Code in Java
Inheritance is a core concept in Object-Oriented Programming (OOP) that allows a class (subclass or derived class) to inherit properties and behaviors from another class (superclass or base class). This promotes code reuse, hierarchical organization of classes, extensibility and leading to a more efficient and maintainable codebase.

In this code, both the 'Dogs' and 'Cats' classes inherit the 'species' and 'age' attributes from the 'Animals' class, showcasing how inheritance promotes code reusability and hierarchical organization.